increase possible size

+++ db/migrate/20240208212445_create_timeslots.rb
@@ -1,7 +1,7 @@
 class CreateTimeslots < ActiveRecord::Migration[7.1]
   def change
     create_table :timeslots do |t|
-      t.integer :size
+      t.bigint :size
       t.datetime :start_time
       t.datetime :end_time
       t.belongs_to :event, foreign_key: true
